reorganized storage
